+++ examples/beforeExit.js
@@ -1,25 +1,21 @@
-'use strict';
 
-let stopped = false;
+var stopped = false;
 
 function work() {
-    console.log('working');
-    !stopped && setTimeout(work, 200);
-  }
-
-function stop() {
-  stopped = true;
-  console.log('shutting down', stopped);
+  console.log('working');
+  !stopped && setTimeout(work, 1000);
 }
 
-process.once('SIGINT', function() {
-  console.log('SIGINT');
-  stop();
-});   // CTRL-C
-process.once('SIGTERM', function() {
-  console.log('SIGTERM');
-  stop();
-});  // pm2 stop
-process.on('beforeExit', () => console.log('exited cleanly :)'));
+function stop() {
+  console.log('shutting down');
+  stopped = true;
+}
+
+process.once('SIGINT', stop);   // CTRL-C
+
+process.on('beforeExit', function() {
+  console.log('exited cleanly :)');
+  process.exit(0);
+});
 
 work();